restore location of resource and feedback enums
