+++ amundsen_application/static/js/interfaces/Enums.ts
@@ -1,16 +1,3 @@
-export enum SendingState {
-  ERROR = "error",
-  IDLE = "idle",
-  WAITING = "waiting",
-  COMPLETE = "complete"
-};
-
-export enum ResourceType {
-  table = "table",
-  user = "user",
-  dashboard = "dashboard",
-};
-
 export enum UpdateMethod {
   PUT = 'PUT',
   DELETE = 'DELETE',
