+++ amundsen_application/static/js/interfaces/Feedback.ts
@@ -1,0 +1,6 @@
+export enum SendingState {
+  ERROR = "error",
+  IDLE = "idle",
+  WAITING = "waiting",
+  COMPLETE = "complete"
+};
