+++ amundsen_application/static/js/interfaces/index.ts
@@ -1,5 +1,6 @@
 export * from './Announcements';
 export * from './Enums';
+export * from './Feedback';
 export * from './Resources';
 export * from './Search';
 export * from './TableMetadata';
